# src/swagger/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a server API Documentation for our Green Team 😎 \n Feel free to CRUD your leaves \n For any question feel free to ask me on Microsoft Teams or contact me on Telegram ◀️ 😌 \n our server is running on `https://green-team001.herokuapp.com/api-docks ` ",
    "version": "1.0.0",
    "title": "Swagger Trello"
  },
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/comment/{id}": {
      "get": {
        "tags": [
          "comment"
        ],
        "summary": "Find Comment by ID",
        "description": "Returns a single comment",
        "operationId": "getCommentById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of leaf to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "example": {
                "comment_id": 14,
                "content": "Great Eco Station",
                "createdAt": "2021-10-11T13:34:17.990Z"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      },
      "patch": {
        "tags": [
          "comment"
        ],
        "summary": "Updates a comment in db with form data",
        "description": "",
        "operationId": "updateComment",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of comment that needs to be update",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "description": "property that needs to be updated",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "example": {
                "content": "Great Eco Station, Only they have short working hours"
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Comment not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "delete": {
        "tags": [
          "comment"
        ],
        "summary": "Deletes a comment",
        "description": "",
        "operationId": "deleteComment",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Comment id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Comment not found"
          }
        }
      }
    }
  }
}